--- packages/read/src/providers/index.ts
@@ -1,2 +1,0 @@
-export { FileSystem } from './file-system';
-export { GitHub } from './github';
